+++ spec/truffle/capi/ext/unimplemented_spec.c
@@ -14,16 +14,16 @@
 extern "C" {
 #endif
 
-static VALUE rb_tr_error_spec_not_implemented(VALUE self, VALUE str) {
-  // One of the not implemented ruby.c functions using rb_tr_error()
+static VALUE unimplemented_spec_not_implemented(VALUE self, VALUE str) {
+  // One of the functions not implemented in ruby.c
   rb_str_shared_replace(str, str);
   return Qnil;
 }
 
-void Init_rb_tr_error_spec(void) {
+void Init_unimplemented_spec(void) {
   VALUE cls;
   cls = rb_define_class("CApiRbTrErrorSpecs", rb_cObject);
-  rb_define_method(cls, "not_implemented_function", rb_tr_error_spec_not_implemented, 1);
+  rb_define_method(cls, "not_implemented_function", unimplemented_spec_not_implemented, 1);
 }
 
 #ifdef __cplusplus