+++ spec/truffle/capi/unimplemented_spec.rb
@@ -8,7 +8,7 @@
 
 require_relative '../../ruby/optional/capi/spec_helper'
 
-load_extension("rb_tr_error")
+load_extension("unimplemented")
 
 describe "Unimplemented functions in the C-API" do
   before :each do
@@ -18,6 +18,6 @@ describe "Unimplemented functions in the C-API" do
   it "raise a useful RuntimeError including the function name" do
     -> {
       @s.not_implemented_function("foo")
-    }.should raise_error(RuntimeError, "rb_str_shared_replace not implemented")
+    }.should raise_error(RuntimeError, /rb_str_shared_replace cannot be found/)
   end
 end